+++ Assets/com.mutant.core/Runtime/Events/EventBus.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
+using Mutant.Core.Diagnostics;
 
 namespace Mutant.Core.Events
 {
@@ -18,6 +19,8 @@ namespace Mutant.Core.Events
 				Handlers[type] = Delegate.Combine(existing, handler);
 			else
 				Handlers[type] = handler;
+
+			CoreRecorder.Record("EventBus", $"Subscribe<{type.Name}>");
 		}
 
 		public static void Unsubscribe<T>(Action<T> handler)
@@ -36,6 +39,8 @@ namespace Mutant.Core.Events
 				Handlers.Remove(type);
 			else
 				Handlers[type] = current;
+
+			CoreRecorder.Record("EventBus", $"Unsubscribe<{type.Name}>");
 		}
 
 		public static void Publish<T>(T evt)
@@ -43,12 +48,16 @@ namespace Mutant.Core.Events
 			Type type = typeof(T);
 
 			if (Handlers.TryGetValue(type, out Delegate del) && del is Action<T> callback)
+			{
 				callback.Invoke(evt);
+				CoreRecorder.Record("EventBus", $"Publish<{type.Name}>");
+			}
 		}
 
 		public static void Clear()
 		{
 			Handlers.Clear();
+			CoreRecorder.Record("EventBus", "Clear");
 		}
 	}
 }